# backend/src/main/resources/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "contact": {
      "email": "info@springonly.com",
      "name": "Equipo de Desarrollo de Spring Only"
    },
    "description": "API para información y gestión de créditos simples.",
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "title": "Credit API REST",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://backend.springonly.com",
      "description": "Servidor Spring Only"
    }
  ],
  "components": {
    "schemas": {
...
      "CreateLoanRequest": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string"
          },
          "currencyId": {
            "type": "string"
          },
          "principalAmount": {
            "type": "number"
          },
          "interestRate": {
            "type": "number"
          },
          "loanDisbursementDate": {
            "type": "string",
            "format": "date",
            "examples": [
              "2022-03-10"
            ]
          },
          "numberOfMonthlyPayments": {
            "type": "integer",
            "format": "int16"
          }
        }
      },
      "CreateLoanResponse": {
        "type": "object",
        "properties": {
          "loanId": {
            "type": "integer",
            "format": "int32"
          },
          "customerId": {
            "type": "string"
          },
          "currencyId": {
            "type": "string"
          },
          "principalAmount": {
            "type": "number"
          },
          "interestRate": {
            "type": "number"
          },
          "loanDisbursementDate": {
            "type": "string",
            "format": "date",
            "examples": [
              "2022-03-10"
            ]
          },
          "numberOfMonthlyPayments": {
            "type": "integer",
            "format": "int16"
          },
          "loanStateId": {
            "type": "string"
          },
          "writtenAt": {
            "type": "string",
            "format": "date-time",
            "examples": [
              "2022-03-10T12:15:50-04:00"
            ]
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        }
      },
...
      "GetLoanByIdResponse": {
        "type": "object",
        "properties": {
          "loanId": {
            "type": "integer",
            "format": "int32"
          },
          "customerId": {
            "type": "string"
          },
          "currencyId": {
            "type": "string"
          },
          "principalAmount": {
            "type": "number"
          },
          "interestRate": {
            "type": "number"
          },
          "loanDisbursementDate": {
            "type": "string",
            "format": "date",
            "examples": [
              "2022-03-10"
            ]
          },
          "numberOfMonthlyPayments": {
            "type": "integer",
            "format": "int16"
          },
          "loanStateId": {
            "type": "string"
          },
          "writtenAt": {
            "type": "string",
            "format": "date-time",
            "examples": [
              "2022-03-10T12:15:50-04:00"
            ]
          }
        }
      },
...
      "UpdateLoanRequest": {
        "type": "object",
        "properties": {
          "loanStateId": {
            "type": "string"
          }
        }
      },
      "UpdateLoanResponse": {
        "type": "object",
        "properties": {
          "loanId": {
            "type": "integer",
            "format": "int32"
          },
          "customerId": {
            "type": "string"
          },
          "currencyId": {
            "type": "string"
          },
          "principalAmount": {
            "type": "number"
          },
          "interestRate": {
            "type": "number"
          },
          "loanDisbursementDate": {
            "type": "string",
            "format": "date",
            "examples": [
              "2022-03-10"
            ]
          },
          "numberOfMonthlyPayments": {
            "type": "integer",
            "format": "int16"
          },
          "loanStateId": {
            "type": "string"
          },
          "writtenAt": {
            "type": "string",
            "format": "date-time",
            "examples": [
              "2022-03-10T12:15:50-04:00"
            ]
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/customers/{customerId}/loans": {
      "parameters": [
        {
          "name": "X-RelationshipManager-Id",
          "in": "header",
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "summary": "List loans by customer",
        "description": "List loans associated with the specified customerId.",
        "operationId": "listLoansByCustomerId",
        "parameters": [
          {
            "name": "customerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Loans list for customer",
            "content": {
              "application/json": {
                "examples": {
                  "Customer loans": {
                    "summary": "A list of loans for a customer",
                    "value": [
                      {
                        "loanId": 1,
                        "customerId": "CUST12345",
                        "currencyId": "PEN",
                        "principalAmount": 10000.0,
                        "interestRate": 5.5,
                        "loanDisbursementDate": "2023-01-01",
                        "numberOfMonthlyPayments": 12,
                        "loanStateId": "VIGENTE",
                        "writtenAt": "2025-09-28T11:15:00-05:00"
                      },
                      {
                        "loanId": 2,
                        "customerId": "CUST12345",
                        "currencyId": "USD",
                        "principalAmount": 20000.0,
                        "interestRate": 6.0,
                        "loanDisbursementDate": "2023-02-01",
                        "numberOfMonthlyPayments": 24,
                        "loanStateId": "DESAPROBADO",
                        "writtenAt": "2025-08-30T03:49:02-05:00"
                      }
                    ]
                  }
                },
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetLoanByIdResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No loans found for customer",
            "content": {
              "application/json": {
                "examples": {
                  "No loans found": {
                    "summary": "An example of a customer with no loans",
                    "value": {
                      "message": "No loans found for customer",
                      "code": "ERR404"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Customer Resource"
        ]
      }
    },
    "/api/loans": {
      "parameters": [
        {
          "name": "X-RelationshipManager-Id",
          "in": "header",
          "schema": {
            "type": "string"
          }
        }
      ],
      "post": {
        "summary": "Create a new loan",
        "description": "Creates a loan for a customer. Requires a CreateLoanRequest payload with the necessary loan data.",
        "operationId": "createLoan",
        "requestBody": {
          "description": "CreateLoanRequest payload",
          "content": {
            "application/json": {
              "examples": {
                "Create loan example": {
                  "summary": "Request to create a loan",
                  "value": {
                    "customerId": "CUST12345",
                    "currencyId": "PEN",
                    "principalAmount": 10000.0,
                    "interestRate": 5.5,
                    "loanDisbursementDate": "2025-09-01",
                    "numberOfMonthlyPayments": 36
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/CreateLoanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Loan created successfully",
            "content": {
              "application/json": {
                "examples": {
                  "CreateLoanResponse example": {
                    "summary": "Created loan representation",
                    "value": {
                      "loanId": 12345,
                      "customerId": "CUST12345",
                      "currencyId": "PEN",
                      "principalAmount": 10000.0,
                      "interestRate": 5.5,
                      "loanDisbursementDate": "2025-09-01",
                      "numberOfMonthlyPayments": 36,
                      "loanStateId": "EN EVALUACION",
                      "writtenAt": "2025-10-07T22:57:09-05:00"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/CreateLoanResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "examples": {
                  "CreateLoan error": {
                    "summary": "Invalid input example",
                    "value": {
                      "message": "principalAmount must be greater than 0",
                      "code": "INVALID_REQUEST"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Loan Resource"
        ]
      }
    },
    "/api/loans/{loanId}": {
      "parameters": [
        {
          "name": "X-RelationshipManager-Id",
          "in": "header",
          "schema": {
            "type": "string"
          }
        }
      ],
      "patch": {
        "summary": "Update an existing loan",
        "description": "Updates loan fields for the given loanId using UpdateLoanRequest.",
        "operationId": "updateLoan",
        "parameters": [
          {
            "name": "loanId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "UpdateLoanRequest payload",
          "content": {
            "application/json": {
              "examples": {
                "Update loan example": {
                  "summary": "Request to update a loan state",
                  "value": {
                    "loanStateId": "VIGENTE"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/UpdateLoanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Loan updated successfully",
            "content": {
              "application/json": {
                "examples": {
                  "UpdateLoanResponse example": {
                    "summary": "Updated loan representation",
                    "value": {
                      "loanId": 12345,
                      "customerId": "CUST12345",
                      "currencyId": "USD",
                      "principalAmount": 10000.0,
                      "interestRate": 5.5,
                      "loanDisbursementDate": "2025-09-01",
                      "numberOfMonthlyPayments": 36,
                      "loanStateId": "VIGENTE",
                      "writtenAt": "2025-10-07T23:05:32-05:00"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/UpdateLoanResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "examples": {
                  "UpdateLoan error": {
                    "summary": "Invalid update example",
                    "value": {
                      "message": "Invalid update data",
                      "code": "INVALID_UPDATE"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Loan not found",
            "content": {
              "application/json": {
                "examples": {
                  "UpdateLoan not found": {
                    "summary": "Loan to update was not found",
                    "value": {
                      "message": "El préstamo indicado no existe",
                      "code": "LN002"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Loan Resource"
        ]
      },
      "get": {
        "summary": "Get loan by ID",
        "description": "Retrieves the loan details for the provided loanId.",
        "operationId": "getLoanById",
        "parameters": [
          {
            "name": "loanId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Loan retrieved successfully",
            "content": {
              "application/json": {
                "examples": {
                  "GetLoanByIdResponse example": {
                    "summary": "A loan retrieved by id",
                    "value": {
                      "loanId": 12345,
                      "customerId": "CUST12345",
                      "currencyId": "PEN",
                      "principalAmount": 10000.0,
                      "interestRate": 5.5,
                      "loanDisbursementDate": "2025-09-01",
                      "numberOfMonthlyPayments": 36,
                      "loanStateId": "VIGENTE",
                      "writtenAt": "2025-09-16T15:45:00-05:00"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/GetLoanByIdResponse"
                }
              }
            }
          },
          "404": {
            "description": "Loan not found",
            "content": {
              "application/json": {
                "examples": {
                  "GetLoanById error": {
                    "summary": "Loan not found example",
                    "value": {
                      "message": "El préstamo indicado no existe",
                      "code": "LN003"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Loan Resource"
        ]
      }
    },
...
  }
}